Copilot chat panel for VS Code (#2151)
This adds a panel to VS Code that provides Copilot chat functionality. 

We define a set of client-side tools to be invoked by the model. These
tools interface with the user's Azure Quantum workspace, providing a way
to submit jobs, get information about the user's jobs and available
targets etc.

The panel isn't yet enabled by default. It requires a special
configuration to be defined in VS Code settings.