Distance coefficient power parameter in QEC scheme (#2212)
This PR extends the formula template for QEC logical error rate formulas
from:

$$ a \left(\frac{p}{p_{\mathrm{th}}}\right)^{\frac{d+1}{2}} $$

to

$$ a d^k \left(\frac{p}{p_{\mathrm{th}}}\right)^{\frac{d+1}{2}} $$

where $k$ is the new distance coefficient power discussed in
[arXiv:2411.10406, Section III.B, page
25](https://arxiv.org/pdf/2411.10406#page=25).

For $k = 0$, the default, the formula remains unchanged compared to the
current one. However, changing $k$ to other values allows to model a
different error correction behavior, as motivated by the paper.

The parameter is called `distance_coefficient_power` or
`distanceCoefficientPower` in the JSON parameter. It is also added to
the Python RE parameter utility structure (together with
`max_code_distance` which was absent so far).