Fixed coefficients in qubit kata explanation. (#2225)
Indeed, atan2(y, x) is present in many languages. We also have
ArcTan2(y, x) in Q#, and it takes y as the first argument as expected.
The text describes an example of a phase of a complex number $\frac{1 -
i}{2}$. Here x=1/2 and y=-1/2. Previous text had these coefficients
switched. Now it is fixed.
Also changed wording from "the phase of |0>" to "the phase of |0>
amplitude", which seems to be a better wording as the paragraph talks
about phases of complex numbers.

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>