Decompose measurement in QIR Base profile processing (#2230)
This change simplifies the decomposition of `M` and `Measure` in the
stdlib to no longer conditionally compile to different versions based on
qubit reuse capability, instead offloading that to QIR processing, which
already has similar logic.

One effect of this is simulation and circuit generation will match user
expectation and patterns in the Q# code rather than more closely showing
what occurs on hardware.