Fix Test Explorer issues (#2291)
This PR fixes two issues.

Firstly, Q# code used in chat contexts was showing up in the Test
Explorer as additional documents (and possibly other locations). It
seems to use a few schemes all starting with 'chat', so excluding those.

Secondly, the location given for failing tests was often wrong, and
falling back to the first 'related' location span when we're also using
that for the document uri anyway resolves that issue. (And seems like
the correct thing to do anyway).