Circuit Editor (#2238)
This PR adds a custom editor view in the VS Code extension for files
with the .qsc extension. The editor makes use of the same library code
that is used to draw circuit diagrams elsewhere in the VS Code
extension. The editor provides basic drag-and-drop editing of circuits
with a simple gate set, including the ability to have controlled gates,
and gates with arguments.
The circuit files are included in Q# projects and can be called into
from Q# as if they were Q# operations. This is done by generating Q#
files for circuit files when compiling.

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>