Add lint groups to Q# (#2103)
This PR adds the necessary infrastructure to define lint groups. It also
adds the `deprecation` lint group, which allows the user to enable all
deprecation lints at the same time, which are set to `allow` by default.

A lint group can be added to the `qsharp.json` manifest as follows:

```json
{
   "lints": [
      { "group": "deprecations", "level": "warn" }
   ]
}
```