Capitalize the Rotation gate names in trace circuits (#2328)
The rotation gates created by the circuit builder were getting
lower-case names: `rx`, `rxx` for example. These are inconsistent with
the other gate names like `H` and `X`. It is also inconsistent with the
Q# equivalent operations `Rx(theta, q)`.

This PR changes the output of the circuit builder to have upper-case
names for the rotation gates, making them consistent with the other gate
names and the Q# operations.