Copilot tools for run, estimate, circuit (#2380)
Adds new `LanguageModelTool`s for running a Q# program locally, resource
estimation, and circuit diagrams.

Just as when they're explicitly invoked, these commands bring up the
webview panels to show results. The results are also added to the chat
context so that the language model can talk about and/or interpret them.

Example:



![image](https://github.com/user-attachments/assets/7487fba6-2caf-4533-bbcc-c09a99c58f13)

---------

Co-authored-by: Bill Ticehurst <billti@microsoft.com>