Limit inferred type sizes (#2382)
This change introduces a notion of type "size" as a way to detect and
avoid degenerate type inference cases. The examples found by the fuzzer
can create large, complex types that result in hangs during type
inference. Much like the type inference recursion limit, this size limit
is a heuristic that is unlikely to be hit during normal use of the
language but prevents crashes or hangs during extreme usage such as with
generated code that incurs artificially large or deeply nested types.
Fixes #2369