Updates some of the CSS colors for Circuits (#2409)
Made some suggested changes to the color schemes of some of the details
in the Circuit Editor:
- The disabled "OK" button now uses the same colors as other buttons and
just uses an opacity to indicate its unavailability.

![image](https://github.com/user-attachments/assets/e2cc2fc8-59a5-42cd-9fbb-56bccc37b30b)
- The default color for button boarders is now transparent so if a theme
doesn't define a border color, boarders do not appear.
- The argument on gates will remain underlined and have the "clickable"
cursor on hover, but the color will not be hyperlink blue as that did
not mesh well with certain themes. The color is now the same as the
regular text color.

![image](https://github.com/user-attachments/assets/54cb6d26-eba7-45c7-82fe-34ff55a626e5)