Support Array Update Syntax (#2414)
This PR supports array syntax as described in #2378.
```qsharp
arr[0][1][2] = 3
```
gets transformed in the HIR to
```qsharp
arr w/= 0 <- (arr[0] w/ 1 <- (arr[0][1] w/ 2 <- 3))
```
This allows user to not have to use the `w/=` syntax to update arrays.
This also supports range expressions for the index. It is also careful
to make sure the index expressions are only evaluated once and in order.

This PR also adds lints for discouraging the use of `w/=` and `w/`
syntax, with a code action for the `w/=` to convert it to the new syntax
here. Both of those lints are set to `Allow` by default, but can be
changed to `Warn` in the future.

This also supports compound assignment operators, so you can do
`arr[0][1][2] += 3` and it will be converted to the new syntax.

---------

Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>