Allow implicit cast to non-const in binary exprs. (#2430)
This PR fixes an issue in casting where the following code was treated
as an error.
```
int x = 5;
int y = 2 * x;
```

The issue was that `2` is a constant, and the casting system wasn't
allowing `x`, of type `int`, which isn't a constant to be cast to the
type of `2`, which is `const int`. The fix consists in allowing the type
of the binary expression as a whole to be demoted to a non-const type.