Add sample notebooks (#2437)
This PR adds sample Python integration and resource estimation notebooks
for OpenQASM. There is more here than just those changes as existing
external issues require workarounds.

- Qiskit and existing OpenQASM code incorrectly specify 0 for angles
which isn't allowed in the spec. So this PR adds the ability to cast
literals to angles with a value of 0 for compatibility. This also
removes some casting calls as we can now directly create the angle from
the literal.
- There is a util bitstring function which allows the user to specify
that they want the output of simulation to be converted into a bitstring
if any of the outputs are bit registers. This matches what Qiskit output
would be like. For the moment the code defaults to Q#'s typically output
rendering.
- The old parser couldn't handle gates, but with the new parser we don't
have to fully transpile into the Q# canonical gates. This changes the
resource estimation numbers.