Fix bit shifts with bit literals on lhs (#2450)
The following program was failing to compile because the `1` on the lhs
of the `<<` expression was being lowered as an `int`, and the qasm spec
does not define bit shifts for ints. The fix consists in lowering the
literal as a `uint` instead, which is allowed by the spec.

```
const int a = 1 << 5;
def const_evaluation_context() {
    int b = a;
}
```