Strengthen check for inference type loops (#2448)
This change updats the check we have in type inference to ensure an
inference type is not bound to another type that includes itself. The
previous check worked for simple cases where the target type directly
contained the inference type, but fails to catch recursion where the
target type contained inference types that indicrectly pointed to the
original inference type through the current type inference solution.
With this check in place, we can now explicitly give an error when a
recursive type is detected.
Fixes #2426