Allow to catch and propagate errors in all resource estimation models (#2457)
This changes return types for methods in the `Overhead` and
`FactoryBuilder` trait from `T` to `Result<T,
resource_estimator::estimates::Error>`.

No implementation of this `Overhead` and `FactoryBuilder` trait returns
an error; therefore, the behavior of RE for the default system
architecture does not change. However, when implementing the traits to
extend the resource estimator, the new signature allows more flexibility
and allows users of the implementation to see errors.