Support Running Projects from Circuit Files (#2455)
Adds the VS Code Run button to circuit files so that the Q# project they
are a part of can be run or debugged while looking at a circuit file.
~~Also adds a command that the Run button can use for running just the
circuit file and producing a dump machine for the circuit. This new
functionality is available even if the circuit is not part of a Q#
project.~~ The Run Circuit File command has been removed and will be
part of a separate future PR.