Clarified search goal in the Grover sample comments (#2509)
Updated comments in the Grover search sample to clarify what we are
searching for. In this particular case we are looking for the marked
state |01010⟩. I still think we should keep this sample simple by
looking for a specific state represented by a hard-coded operation
rather than somehow providing arbitrary marked state as a parameter.

---------

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>