Add support for builtin functions (#2508)
This PR adds the necessary infrastructure to support builtin functions
in the lowerer and prepares the groundwork for providing language
service support for const evaluated expressions. These changes are
related because builtin function calls are a kind of const evaluated
expression.

The PR can be understood in two areas:
1. The part that prepares the groundwork for providing language service
support for const evaluated expressions, which includes two changes:
const expressions are now eagerly evaluated in the lowerer; and the
`Expr` type is augmented with the result of const evaluating the
expression, instead of just storing the final value in the AST, this
will provide the language service with the necessary span and type
information for its hover, jump to definition, and type hint features.
2. The part that adds the necessary infrastructure to support builtin
functions in the lowerer: this changes include adding a polymorphic
dispatch mechanism for QASM bultins and wiring this mechanism to the
`Lowerer::lower_function_call_expr` method.