Prevent Stepping into Circuit Files during Debug (#2480)
This adds a de-spanning pass to the compiler to remove the spans for the
contents of operations defined via circuit files. The operation
declarations themselves retain their spans to support LS features, but
the statements in the body of the operations are de-spanned. This
prevents the debugger from stepping into the circuit files.