Add EC trait function to adjust code parameter after selection (#2516)
This adds an EC trait function to provide the possibility to adjust the
code parameter after it has been computed. This is required for special
EC functionality through the extensibility API. The default
implementation of the trait function returns the original code parameter
and the EC implementation inside the QDK does not change that function.