Remove implicit modifier handling during OpenQASM lowering (#2529)
In the original implementation of the OpenQASM compiler, gates with
implicit modifiers needed to be specially handled by redefining them as
their unmodiefied variants and collecting the modifier. For example,
`ch` would be converted to `h` with a `ctrl(1)` modifier and would be
used to build up the gate call with the additional explicit modifiers.

With the introduction of the `Std.OpenQASM` library, we can now call
those gates directly without needing to decompose them. This also cleans
up error messages as any errors would have been associated with the
unmodified gate name prior.