Classical types indexing in OpenQASM (#2564)
This PR adds classical type indexing support to our OpenQASM compiler
and fixes two bugs.

## Highlights
The major challenge of this PR was supporting classical type indexing in
the left-hand side of assignment expressions. The idea to make it work
is compiling
```
int[4] a = 0b1111;
a[1:][1:] = "01";
```
to this
```
mutable a = 15;
a = {
    mutable bitarray = Std.OpenQASM.Convert.IntAsResultArrayBE(a, 4);
    bitarray[1...][1...] = [Zero, One];
    Std.OpenQASM.Convert.ResultArrayAsIntBE(bitarray)
};
```
so that we can leverage Q#'s array semantics.
 
The indexing infrastructure is now simpler. Instead of having different
codepaths for indexed identifiers and indexed expressions, now we just
have a codepath for indexed expressions. (The codepath for indexed
identifiers was redundant).

Multiple indexing is now expressed as nested IndexedExpr structs with a
single index each, instead of a single `IndexedExpr` with a list of
indices. This makes thinking about indexing simpler, since now the
utility functions to handle indexing just need to handle a single index.
It also makes compilation simpler, since this is how the Q# AST is
structured.

## Bug fixes:
1. Consecutive slices into an array now act on the same dimension,
instead of acting on consecutive dimensions.
2. `AngleAsResultArrayBE` is now correctly exported in the `Angle.qs`
file instead of `AngleAsResultArray`.