Upgrade the rust version to 1.88 (#2583)
This PR upgrades the rust toolchain version from 1.86 to 1.88 across
multiple workflows and files. To make the upgrade possible, we had to
fix a linker issue and a wasm-config flag that changed in 1.87.

## Linker issue
Starting on rust 1.87, the Std doesn't link advapi32 on windows anymore.
So, C libraries that depend on it, like mimalloc, now need to link it
manually. [Link to the Rust
PR](https://github.com/rust-lang/rust/pull/138233).

## wasm config flag
Rust 1.87 upgraded to LLVM 20, which enables bulk-memory by default.
Therefore, `wasm-opt` needs to be configured to expect this setting.
This behavior is document
[here](https://doc.rust-lang.org/rustc/platform-support/wasm32-unknown-unknown.html#enabled-webassembly-features).

---------

Co-authored-by: Bill Ticehurst <billti@microsoft.com>