Remove wasm-pack dependency (#2587)
wasm-pack has been a bit of a pain, and seems largely unmaintained at
this point (has several hundred issues open and hasn't had a commit in 9
months). We don't use many of its features anyway (such as testing and
packaging for npm).

With this change, we use wasm-bindgen (for the .JS files) and wasm-opt
(to optimize the wasm) directly, instead of via wasm-pack. All wasm-pack
was doing under the covers was installing these tools from the internet
and invoking them, so this just makes their usage more explicit.

With this we can:

- Directly control the versions of the wasm utilities used
- Build wasm on Windows ARM64 (currently this fails in main, but works
with this PR)
- Avoid streaming a .sh script from the internet to install wasm-pack at
build time

This requires both wasm-bindgen and wasm-opt be on the path, but 'python
prereqs.py --install' will add those if needed. (They are handy to have
around if working with wasm anyway).

Open to feedback. It wasn't as simple as I hoped, but should be a
one-off change.