Making MapPauliAxis public with updates to functionality and comments (#2585)
This adds a function to remap Pauli axes `MapPauliAxis` to the new
BlochSphere.qs file. It is based on the previously internal MapPauli and
brings some updates and extensive doc comments. This is useful to apply
operations in different Pauli bases.

For example, rotation around the X axis can be expressed as a rotation
around the Z axis by remapping the Pauli axes so that X axis becomes Z
axis. The following two operations are equivalent:

```
Rx(0.1, q)
```
and
```
within {
    MapPauliAxis(PauliZ, PauliX, q);
} apply {
    Rz(0.1, q);
}
```

This function is used internally in some rotations and state
preparation.
Resolves part 2 of https://github.com/microsoft/qsharp/issues/1104

---------

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>