Add step-point to block exit when debugging
This change adds a new step-point during debugging so that the debugger can stop on each scope exit and see the final values of any variables local to that scope. Like similar debugging features, these new nodes are not added to the graph during non-debug execution, such as evaluating Q# code from Python.