Changelog View Added (#2576)
Adds a changelog view that can be accessed via a new command: "QDK: Show
Changelog". The view is populated by the contents of the CHANGELOG.md
file, and the user is given a notification of a new changelog when the
CHANGELOG_VERSION constant defined in changelog.ts is updated. The user
can opt out of these notifications after the first one, and can opt back
in by changing the new Q# setting.

The CHANGELOG_VERSION is currently set to `undefined` which means that
the changelog page will not automatically appear for users. The first
time users will be notified of a new changelog page will be when the
constant is manually updated to a new value.

Manually Validated the following scenarios:
- With setting = false:
  - CHANGELOG_VERSION = undefined : no popup
- CHANGELOG_VERSION = undefined -> "defined" : no popup but changelog
automatically shows
  - CHANGELOG_VERSION = "defined" : no popup
  - CHANGELOG_VERSION = "defined" -> "defined2" : popup
  - CHANGELOG_VERSION = "defined2" : no popup
- CHANGELOG_VERSION = "defined2" -> undefined : popup (this is
unrealistic scenario)

- With setting = true: (no popup for all cases)
  - CHANGELOG_VERSION = undefined
  - CHANGELOG_VERSION = undefined -> "defined"
  - CHANGELOG_VERSION = "defined"
  - CHANGELOG_VERSION = "defined" -> "defined2"
  - CHANGELOG_VERSION = "defined2"
- CHANGELOG_VERSION = "defined2" -> undefined (this is unrealistic
scenario)

Addresses: #2497