Separate lowerer and compiler errors (#2603)
This PR moves errors from the lowerer to the compiler for errors that
are strictly associated with the QDK's unsupported OpenQASM features. As
as side effect:
- Unified parsed string handling in `Arc`'s
- Removed dead code
- Added `cal` and `defcal` bodies to parsed and semantic ASTs
- Moved Unsupported errors from lowerer to compiler

There is one bit that is a little annoying. When fixing the
function/callable/def type to track the args and return type, we now
have a scenario where the symbol doesn't have the correct type
information. Before, we pulled the return type from the def stmt instead
of the symbol. Now, we pull the return type from the symbol ty, but in
cases where this shadowing happens, we set the ty to `Err` since we
don't have the correct type. We can either put the return ty back into
the stmt (duplicating info or in this case providing accurate info) or
leave it as is. I don't think there is a correct answer as the code is
incorrect. Q# allows for shadowing and if the correct type is provided,
the errors aren't correct for usages of the function.