Added telemetry to changelog prompt (#2604)
Added two telemetry events, `ChangelogPromptStart` and
`ChangelogPromptEnd` for tracking the use of the changelog prompt.
`ChangelogPromptStart` will fire when we show the popup and has a
`qdkVersion` field for what version of the QDK the popup is being shown
for. `ChangelogPromptEnd` will fire when the user clicks one of the
buttons on the changelog popup and has an `action` field with possible
values `"showChangelog"` or `"suppressChangelog"` for discerning which
button was pressed. Both events have and `associationId` field to
correlate instances of the two.

Also removed the restriction that prevents the popup from appearing on a
fresh install, as we want the popup to appear in that scenario now.