Fix endianness bug when indexing into OpenQASM classical types (#2600)
The following list of conclusions gathered from the OpenQASM spec will
be useful while reviewing this PR. You can find this information in the
official [OpenQASM spec](https://openqasm.com/versions/3.1/index.html).

1. Bitstring Literals are written in big-endian order.
The spec says: "For convenience, classical registers can be assigned a
text string containing zeros and ones of the same length as the size of
the register. It is interpreted to assign each bit of the register to
corresponding value 0 or 1 in the string, where the least-significant
bit is on the right."

2. Bit indexing of bit arrays literals is in big-endian order.
   The spec says: "classical registers are static arrays of bits."
So, when doing bitarray[0] we are accessing the first element of that
"static array", the leftmost element, which according to (1) is the most
significant bit. Therefore, indexing of bitarrays is in big-endian
order.

3. Bit indexing of int, uint and angle is in little-endian order.
The spec has [this
example](https://openqasm.com/versions/3.0/language/types.html#classical-value-bit-slicing)
where they hint that given `int[32] myInt = 15;`, `myInt[0]` evaluates
to 1, that means that the least significant bit is stored first, and
indexing of int, uint, and angle is in little-endian order.

4. We index classical types by first casting them to bitarrays, which
are in big-endian order, therefore when doing `myInt[0]`, we need to do
`bit[32](myInt)[31]`. Note that we transformed the index `0` into `32 -
0 - 1 = 31`, effectively changing the endianness of the indexing
operation.