Add command to download raw results text (#2609)
A recent PR changed it so that whenever you click the download icon (or
run the "Download Azure Quantum job results" command), instead of the
previous behavior where it would always open the raw results file, it
now shows a histogram if possible.

However it can sometimes still be useful to fetch the raw results data,
so I've added a specific command for that. (I didn't add an additional
icon for it in the tree view as I expect the use-case is rare and I
didn't want to clutter the UX).

With this change, to fetch the raw text of the job results:

1. Select the completed job in the tree view as usual
2. Press F1 to bring up the command palette, and type "job results" or
similar to filter to the command.
3. Select the "... as text" option shown below to get the old behavior
of opening the results file as text in a new editor tab.

<img width="439" height="107" alt="image"
src="https://github.com/user-attachments/assets/29d26cb0-7c0d-4a8e-bb5a-4760ac7572cc"
/>