QIR Profiles Selection via Project Field or EntryPoint argument (#2591)
This removes the workspace setting for the QIR profile and adds a
setting for this for Q# project files instead, i.e. `"targetProfile":
"base"`. Single-file Q# files are supported by adding an argument to the
@EntryPoint attribute where a user will specify what profile the file is
meant to compile with, i.e. `@EntryPoint(Adaptive_RI)`, or uses
`Unrestricted` as a default.