QASM Profile Pragma (#2612)
Adds the `#pragma qdk.qir.profile` pragma to our QASM compiler. This
pragma is used to specify the profile to compile the QASM file with,
i.e. `#pragma qdk.qir.profile Base` to have the file compile with the
`Base` profile. QASM files are always compiled with the profile
specified by their source with this pragma. If no pragma is provided,
the compilation will default to `Unrestricted`.

The PR is built on top of and relies on #2591