Allow recursive calls to operations that return `Unit` (#2654)
This change updates RCA to allow a subset of cyclic operation calls.
These operation calls are allowed as long as the operation returns
`Unit`. The change involves:
- updating the cyclic callable check to avoid populating the compute
properties for cyclic operations that return `Unit`
- updating the core analysis logic to check callee expressions for
cycles in the active context, which are treated as "unresolved" to fall
back to partial evaluation capabilities checks.

Some new test cases confirm the expected behavior for the corner cases,
as well as updates to existing test cases to account for the change that
adds `CallToUnresolvedCallee` capability.