OpenQASM language service: add rename/definition/references, semantic passes, and VS Code wiring (#2656)
## Summary of changes

- Language service (OpenQASM)
- Added OpenQASM support for prepare-rename, rename, go-to-definition,
and find-references.
- Added tests for OpenQASM rename. Other functionalities are
transitively tested as they all use the same code underneath
- OpenQASM semantic layer (compiler)
- Implemented semantic passes and visitors to support symbol lookup and
reference finding (SymbolFinder, ReferenceFinder).
- Fixed and standardized source span tracking (notably for casts)
locations.
- Introduced new semantic AST field indicating whether casts were
implicit or explicit
- VS Code extension
- Registered providers for OpenQASM: completion, definition, references,
and rename.
- In VS Code, OpenQASM files now support:
- Rename symbol (F2) for user-defined identifiers, gates, functions,
etc; built-ins are not renameable.
  - Go to Definition (F12) and Find All References (Shift+F12).

### Notes

- Expressions in type designators are not support with the LS features.
We need a follow up PR to add the support as it will be a significant
set of changes to support this. There is an ignored test for this
scenario for the follow up PR.
- The semantic span fixes (e.g., casts) were required to make
rename/definition locations correct, but it turned out that we couldn't
reuse the Q# AST based LS features, so they are now correct, but aren't
used for LS features.