Lower duration and stretch (#2611)
This PR attempts to lower duration, and stretch values in OpenQASM to
try to make them spec 'compliant'. The duration and stretch types are
supposed to be compile-time constant which would imply they are actually
const decls but not declared in const decl statements.

This is complicated by the `durationof` which is supposed to return a
duration. The `durationof` call will accumulate the durations of boxes,
gate calls, and delays.

**NOTE**
In draft while getting spec clarification. Where we are at:
- Stetches are always const. Default value is `0`. 
- Durations can be but aren't required to be const
- `durationof` is problematic. No real way calculate duration without
defcals. We may need two different calls in the spec. One where it is
const and other call where it returns a non-const duration.