Add missing UDT tag in Python - Q# interop layer (#2677)
There was an edge case in which passing a Python complex number to a Q#
identity function, and getting that output back to Python would result
in a tuple and not in a Python complex number. This was due to a missing
type tag when constructing the Q# Complex UDT value from the Python
object.

This PR adds that missing type tag.