Add support for array concatenation in OpenQASM (#2676)
This PR adds support for [array concatenation in
OpenQASM](https://openqasm.com/versions/3.0/language/types.html#array-concatenation-and-slicing).

### Notes

Array concatenation does not appear in the Grammar and is underspecified
in the Spec. This implementation is conservative in two senses while we
wait for Spec clarification:
1. We don't allow concatenating arrays of different base types, even if
they differ only in width.
2. We don't allow mixing together arrays, static array references, and
dynamic array references.

---------

Co-authored-by: Ian Davis <Ian.F.Davis@microsoft.com>