Copilot Reads Summaries from DocGen Tool (#2672)
Adds a Copilot tool called `qsharpGetLibraryDescriptions` which gives
Copilot updated descriptions of Q# Library API items, allowing it to
make more accurate use of library items in generating Q# snippets and in
answering user questions about the libraries.
The doc get tool is modified to include a way to generate these
summaries in a JSON that is used by the Copilot tool.