Upload supplemental data for Azure jobs (#2686)
Adds an "upload supplemental data" step to Azure job submission.

This step is currently disabled by default, and can be enabled via user
settings (not expected to be useful today except for testing).

If job submission succeeded, we attempt to generate a circuit diagram.
If successful, the resulting JSON diagram data is uploaded to the
`circuitData` blob in the same storage container as the job input data.